+++ R/Uncorrected_asymptotic_score_CI_2x2.R
@@ -5,7 +5,6 @@
 #' @description Described in Chapter 4 "The 2x2 Table"
 #' @param n the observed counts (a 2x2 matrix)
 #' @param alpha the nominal level, e.g. 0.05 for 95% CIs
-#' @param printresults display results (FALSE = no, TRUE = yes)
 #' @examples
 #' # A case-control study of GADA exposure on IPEX syndrome
 #' # (Lampasona et al., 2013):
@@ -17,9 +16,7 @@
 #' @export
 #' @return A data frame containing lower, upper and point estimates of the
 #' tatistic
-Uncorrected_asymptotic_score_CI_2x2 <- function(
-  n, alpha = 0.05, printresults = TRUE
-) {
+Uncorrected_asymptotic_score_CI_2x2 <- function(n, alpha = 0.05) {
   validateArguments(mget(ls()))
 
   # global n11 n21 n1p n2p alphaglobal limit
@@ -66,19 +63,16 @@ Uncorrected_asymptotic_score_CI_2x2 <- function(
     )$root
   }
 
-  if (printresults) {
-    print(
-      sprintf(
-        paste(
-          "Uncorrected asymptotic score CI: estimate =",
-          "%6.4f (%g%% CI %6.4f to %6.4f)"
-        ),
-        estimate, 100 * (1 - alpha), L, U
+  printresults <- function() {
+    sprintf(
+      paste(
+        "Uncorrected asymptotic score CI: estimate =",
+        "%6.4f (%g%% CI %6.4f to %6.4f)"
       ),
-      quote = FALSE
+      estimate, 100 * (1 - alpha), L, U
     )
   }
 
   res <- data.frame(lower = L, upper = U, estimate = estimate)
-  invisible(res)
+  return(contingencytables_result(res, printresults))
 }