+++ R/Wald_test_1x2.R
@@ -5,7 +5,6 @@
 #' @param X the number of successes
 #' @param n the total number of observations
 #' @param pi0 a given probability
-#' @param printresults display results (0 = no, 1 = yes)
 #' @examples
 #' # The number of 1st order male births (adapted from Singh et al. 2010)
 #' Wald_test_1x2(singh_2010["1st", "X"], singh_2010["1st", "n"], pi0 = 0.1)
@@ -19,7 +18,7 @@
 #' Wald_test_1x2(ligarden_2010["X"], ligarden_2010["n"], pi0 = 0.1)
 #' @export
 #' @return A vector containing the two-sided p-value and the Wald test statistic
-Wald_test_1x2 <- function(X, n, pi0, printresults = TRUE) {
+Wald_test_1x2 <- function(X, n, pi0) {
   validateArguments(mget(ls()))
 
   # Estimate of the binomial probability (pihat)
@@ -39,11 +38,11 @@ Wald_test_1x2 <- function(X, n, pi0, printresults = TRUE) {
   # The two-sided P-value (reference distribution: standard normal)
   P <- 2 * (1 - pnorm(abs(Z), 0, 1))
 
-  if (printresults) {
-    print(sprintf("The Wald test: P = %7.5f, Z = %6.3f", P, Z), quote = FALSE)
+  printresults <- function() {
+    sprintf("The Wald test: P = %7.5f, Z = %6.3f", P, Z)
   }
 
   res <- c(P, Z)
   names(res) <- c("p.value", "statistic")
-  invisible(res)
+  return(contingencytables_result(res, printresults))
 }