+++ R/Woolf_logit_CI_2x2.R
@@ -3,7 +3,6 @@
 #' @description Described in Chapter 4 "The 2x2 Table"
 #' @param n the observed table (a 2x2 matrix)
 #' @param alpha the nominal level, e.g. 0.05 for 95% CIs
-#' @param printresults display results (FALSE = no, TRUE = yes)
 #' @examples
 #' # A case-control study of GADA exposure on IPEX syndrome
 #' # (Lampasona et al., 2013):
@@ -14,7 +13,7 @@
 #'
 #' @export
 #' @return A vector containing lower, upper and point estimates of the statistic
-Woolf_logit_CI_2x2 <- function(n, alpha = 0.05, printresults = TRUE) {
+Woolf_logit_CI_2x2 <- function(n, alpha = 0.05) {
   validateArguments(mget(ls()))
 
   # Estimate of the odds ratio (thetahat)
@@ -36,13 +35,13 @@ Woolf_logit_CI_2x2 <- function(n, alpha = 0.05, printresults = TRUE) {
     U <- Inf
   }
 
-  if (printresults) {
-    print(sprintf(
+  printresults <- function() {
+    sprintf(
       "The Woolf logit CI: estimate = %6.4f (%g%% CI %6.4f to %6.4f)",
       estimate, 100 * (1 - alpha), L, U
-    ), quote = FALSE)
+    )
   }
 
   res <- data.frame(lower = L, upper = U, estimate = estimate)
-  invisible(res)
+  return(contingencytables_result(res, printresults))
 }