+++ R/the_paired_2x2_table_CIs_OR.R
@@ -16,32 +16,32 @@ the_paired_2x2_table_CIs_OR <- function(n, alpha = 0.05) {
   my_sprintf("Interval method                              %i%% CI        log width\n", 100 * (1 - alpha))
   print("--------------------------------------------------------------------", quote = FALSE)
 
-  tmp <- Wald_CI_OR_paired_2x2(n, alpha, FALSE)
+  tmp <- Wald_CI_OR_paired_2x2(n, alpha)
   L <- tmp[[1]]
   U <- tmp[[2]]
   my_sprintf("Wald                                   %7.4f to %7.4f   %7.3f\n", L, U, log(U) - log(L))
 
-  tmp <- Wald_CI_OR_Laplace_paired_2x2(n, alpha, FALSE)
+  tmp <- Wald_CI_OR_Laplace_paired_2x2(n, alpha)
   L <- tmp[[1]]
   U <- tmp[[2]]
   my_sprintf("Wald with Laplace adjustment           %7.4f to %7.4f   %7.3f\n", L, U, log(U) - log(L))
 
-  tmp <- Transformed_Wilson_score_CI_paired_2x2(n, alpha, FALSE)
+  tmp <- Transformed_Wilson_score_CI_paired_2x2(n, alpha)
   L <- tmp[[1]]
   U <- tmp[[2]]
   my_sprintf("Transformed Wilson score               %7.4f to %7.4f   %7.3f\n", L, U, log(U) - log(L))
 
-  tmp <- Transformed_Clopper_Pearson_exact_CI_paired_2x2(n, alpha, FALSE)
+  tmp <- Transformed_Clopper_Pearson_exact_CI_paired_2x2(n, alpha)
   L <- tmp[[1]]
   U <- tmp[[2]]
   my_sprintf("Transformed Clopper-Pearson exact      %7.4f to %7.4f  %7.3f\n", L, U, log(U) - log(L))
 
-  tmp <- Transformed_Clopper_Pearson_midP_CI_paired_2x2(n, alpha, FALSE)
+  tmp <- Transformed_Clopper_Pearson_midP_CI_paired_2x2(n, alpha)
   L <- tmp[[1]]
   U <- tmp[[2]]
   my_sprintf("Transformed Clopper-Pearson mid-P      %7.4f to %7.4f   %7.3f\n", L, U, log(U) - log(L))
 
-  tmp <- Transformed_Blaker_exact_CI_paired_2x2(n, alpha, FALSE)
+  tmp <- Transformed_Blaker_exact_CI_paired_2x2(n, alpha)
   L <- tmp[[1]]
   U <- tmp[[2]]
   my_sprintf("Transformed Blaker exact               %7.4f to %7.4f   %7.3f\n", L, U, log(U) - log(L))