+++ tests/testthat/test-ch8.R
@@ -90,7 +90,7 @@ test_that("Chapter 8 functions basically work", {
   )
   expect_output(
     object = print(Wald_CI_OR_paired_2x2(ezra_2010)),
-    regexp = " The Wald CI: estimate = 12.5000 \\(95% CI 2.9608 to 52.7731\\)"
+    regexp = "The Wald CI: estimate = 12.5000 \\(95% CI 2.9608 to 52.7731\\)"
   )
   expect_output(
     object = print(Wald_CI_OR_Laplace_paired_2x2(ezra_2010)),